style: implement intercept route for /photos/[id]
In the previous commit I have created a Modal component, but
this commit introduces the /photos/[id] intercepting route,
and the component for this route implements a modal of its own.

The purpose of this is to see if I could implement a modal myself with
less code than what was implemented in the components/modal/index.tsx
file.